requirements.txt
django==3.0
flake8==3.7.0
pytest-django==3.9.0
six==1.16.0
black==23.9.1
